Add some list projects stuff
