+++ modules/projects/client/controllers/projects.client.controller.js
@@ -1,8 +1,8 @@
 'use strict';
 
 // Projects controller
-angular.module('projects').controller('ProjectsControllers', ['$scope', '$stateParams', '$location', 'Projects',
-  function ($scope, $stateParams, $location, Projects) {
+angular.module('projects').controller('ProjectsController', ['$scope', '$state', '$stateParams', '$location', 'Projects',
+  function ($scope, $state, $stateParams, $location, Projects) {
       //TODO $scope.findTeamMembers
       //   $scope.find = function() {
       //   /* set loader*/
@@ -65,5 +65,53 @@ angular.module('projects').controller('ProjectsControllers', ['$scope', '$stateP
               });
 
       };
+
+    // Remove existing Project
+    $scope.remove = function (project) {
+      if (project) {
+        project.$remove();
+
+        for (var i in $scope.projects) {
+          if ($scope.projects[i] === project) {
+            $scope.projects.splice(i, 1);
+          }
+        }
+      } else {
+        $scope.project.$remove(function () {
+          $location.path('projects');
+        });
+      }
+    };
+
+    // Update existing Project
+    $scope.update = function (isValid) {
+      $scope.error = null;
+
+      if (!isValid) {
+        $scope.$broadcast('show-errors-check-validity', 'projectForm');
+
+        return false;
+      }
+
+      var project = $scope.project;
+
+      project.$update(function () {
+        $location.path('projects/' + project._id);
+      }, function (errorResponse) {
+        $scope.error = errorResponse.data.message;
+      });
+    };
+
+    // Find a list of Projects
+    $scope.find = function () {
+      $scope.projects = Projects.query();
+    };
+
+    // Find existing Project
+    $scope.findOne = function () {
+      $scope.project = Projects.get({
+        projectId: $stateParams.projectId
+      });
+    };
 
   }]);
