+++ modules/projects/server/controllers/projects.server.controller.js
@@ -1,31 +1,60 @@
 'use strict';
 
-/**
- * Module dependencies.
- */
 var path = require('path'),
   mongoose = require('mongoose'),
-  Project = mongoose.model('Project'),
-  errorHandler = require(path.resolve('./modules/core/server/controllers/errors.server.controller'));
+  Project = require('mongoose').model('Project');
 
-
-// Reference modules/articles/server for examples on what can go here
-exports.create = function(req, res) {
+exports.create = function(req, res, next) {
   var project = new Project(req.body);
+  project.save(function (err) {
+    if (err) {
+      return next(err);
+    }
+    else {
+      res.json(project);
+    }
+  });
+};
+exports.list = function(req, res) {
+  Project.find({}, function(err, projects) {
+    if (err) {
+      res.status(400).send(err);
+    }
+    else {
+      res.json(projects);
+    }
+  });
+};
+exports.delete = function(req, res, next) {
+  req.project.remove(function(err) {
+    if (err) {
+      return next(err);
+    }
+    else {
+      res.json(req.project);
+    }
+  });
+};
+exports.read = function(req, res) {
+  res.json(req.project);
+};
 
+exports.projectById = function(req, res, next, id) {
+  Project.findOne({
+      _id: id
+    },
+    function(err, project) {
+      if (err) {
+        return next(err);
+      }
+      else {
+        req.project = project;
+        next();
+      }
+    }
+  );
+};
 
-    exports.create = function (req, res) {
-        var project = new Project(req.body);
-        project.user = req.user;
+exports.update = function (req, res) {
 
-        project.save(function (err) {
-            if (err) {
-                return res.status(400).send({
-                    message: errorHandler.getErrorMessage(err)
-                });
-            } else {
-                res.json(project);
-            }
-        });
-    };
 };
