+++ modules/projects/server/routes/projects.server.routes.js
@@ -5,6 +5,16 @@
  */
 var projects = require('../controllers/projects.server.controller');
 
-module.exports = function (app) {
-  // Reference modules/articles/server for examples on what can go here
+module.exports = function(app) {
+
+  app.route('/api/projects')
+    .post(projects.create)
+    .get(projects.list);
+
+  app.route('/api/projects/:projectId')
+    .get(projects.read)
+    .put(projects.update);
+
+  //
+  // app.param('projectId', projects.projectById);
 };
